Move glog to a FetchContent, Add GoogleTest, Move SmartEnum test into GoogleTest
