+++ src/CMakeLists.txt
@@ -19,6 +19,32 @@ set(CMAKE_MODULE_PATH "${CMAKE_MODULE_PATH};${CMAKE_CURRENT_LIST_DIR}/cmake/modu
 set(CPPCHECK_EXCLUDES "${CMAKE_CURRENT_LIST_DIR}/utility/external;${CMAKE_CURRENT_LIST_DIR}/clue_autogen;${CMAKE_CURRENT_LIST_DIR}/utility/UTILITYlib_autogen;${CMAKE_CURRENT_LIST_DIR}/core/CORElib_autogen;")
 find_package(CppCheck)
 
+include(FetchContent)
+    FetchContent_Declare(
+    googletest
+    GIT_REPOSITORY https://github.com/google/googletest.git
+    GIT_TAG v1.13.0
+)
+# For Windows: Prevent overriding the parent project's compiler/linker settings
+set(gtest_force_shared_crt ON CACHE BOOL "" FORCE)
+FetchContent_MakeAvailable(googletest)
+
+enable_testing()
+include(CTest)
+
+include(FetchContent)
+
+set(WITH_GFLAGS OFF)
+set(WITH_GTEST ON)
+set(WITH_UNWIND ON) # May need to be off if you don't have libunwind or are on Windows
+set(BUILD_TESTING OFF)
+
+FetchContent_Declare(
+    glog
+    GIT_REPOSITORY https://github.com/google/glog.git
+)
+FetchContent_MakeAvailable(glog)
+
 find_package(QT NAMES Qt6 Qt5 COMPONENTS Widgets LinguistTools REQUIRED)
 find_package(Qt${QT_VERSION_MAJOR} COMPONENTS Widgets LinguistTools REQUIRED)
