+++ src/clue_en_US.ts
@@ -5,97 +5,81 @@
     <name>Clue::ConfigureGame</name>
     <message>
         <location filename="configuregame.ui" line="14"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="258"/>
         <source>Dialog</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="45"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="259"/>
         <source>Your Name</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="55"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="260"/>
         <source># of Players</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="72"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="261"/>
         <source>P1 Name</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="82"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="262"/>
         <source>P2 Name</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="92"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="263"/>
         <source>P3 Name</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="102"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="264"/>
         <source>P4 Name</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="112"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="265"/>
         <source>P5 Name</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="122"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="266"/>
         <source>Who&apos;s First?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="139"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="267"/>
         <source>Your Card 0</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="149"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="268"/>
         <source>Your Card 1</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="159"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="269"/>
         <source>Your Card 2</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="169"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="270"/>
         <source>Your Card 3</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="179"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="271"/>
         <source>Your Card 4</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="189"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="272"/>
         <source>Your Card 5</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="configuregame.ui" line="206"/>
-        <location filename="clue_autogen/include/ui_configuregame.h" line="273"/>
         <source># of Cards</source>
         <translation type="unfinished"></translation>
     </message>
@@ -104,123 +88,103 @@
     <name>Clue::MainWindow</name>
     <message>
         <location filename="mainwindow.ui" line="14"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="345"/>
         <source>MainWindow</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="31"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="360"/>
         <source>Enter a Turn</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="48"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="349"/>
         <source>Player&apos;s Turn:</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="69"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="350"/>
         <source>Are they/you making a Suggestion?</source>
         <oldsource>Are they/you making an accusation?</oldsource>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="78"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="351"/>
         <source>Yes</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="91"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="352"/>
         <source>No</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="116"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="353"/>
         <source>What&apos;s the suggestion?</source>
         <oldsource>What&apos;s the Guess?</oldsource>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="131"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="354"/>
         <source>Suspect</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="138"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="355"/>
         <source>Weapon</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="145"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="356"/>
         <source>Room</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="170"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="357"/>
         <source>Who proved the suggestion false?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="191"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="358"/>
         <source>You saw or showed a card?</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="203"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="359"/>
         <source>Card you showed OR card you saw</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="230"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="361"/>
         <source>Known Info</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="245"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="362"/>
         <source>Turns</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="294"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="363"/>
         <source>File</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="301"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="364"/>
         <source>View</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="311"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="346"/>
         <source>Reset</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="316"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="347"/>
         <source>Quit</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
         <location filename="mainwindow.ui" line="327"/>
-        <location filename="clue_autogen/include/ui_mainwindow.h" line="348"/>
         <source>History</source>
         <translation type="unfinished"></translation>
     </message>
